+++ src/CodesResearchTools/Analyzers/Spectrum/ISpectrumAnalyzer.cs
@@ -2,15 +2,16 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Numerics;
     using CodesAbstractions;
     using GfPolynoms;
     using GfPolynoms.GaloisFields;
 
     public interface ISpectrumAnalyzer
     {
-        IReadOnlyDictionary<int, long> Analyze(ICode code, SpectrumAnalyzerOptions options = null);
+        IReadOnlyDictionary<int, BigInteger> Analyze(ICode code, SpectrumAnalyzerOptions options = null);
 
-        IReadOnlyDictionary<int, long> Analyze(
+        IReadOnlyDictionary<int, BigInteger> Analyze(
             GaloisField field,
             int informationWordLength,
             Func<FieldElement[], FieldElement[]> encodingProcedure,
